Remove option log_exception, always log it under our logger.
This unifies the thought process so that everything is done via the
log lines, not these separate config variables. Caller can mute this
module if they want to. Also, it was previously appending directly
to the message but now we can have our logger's name on it so it's
more clear who's reporting what.