Use named column inserts, but remove require_all parameter.
I am yet again bumping into the limits of extract_table_column_map
as it does not use a real parser and does not recognize generated
columns, which causes issues when I use that column list to enforce
an insert.

It probably would be better to move even farther away from extracting
columns and just asking the database instead.

Anyway we get slightly improved ergonomics of insert_filler at the cost
of not being able to enforce all columns within this function.